增删改查，动态更新，自定义类加载器

+++ thirteen-datamation/src/main/java/org/thirteen/datamation/service/impl/DatamationServiceImpl.java
@@ -9,6 +9,7 @@ import org.springframework.stereotype.Service;
 import org.thirteen.datamation.core.criteria.DmCriteria;
 import org.thirteen.datamation.core.criteria.DmQuery;
 import org.thirteen.datamation.core.criteria.DmSpecification;
+import org.thirteen.datamation.core.generate.ClassInfo;
 import org.thirteen.datamation.core.generate.repository.BaseRepository;
 import org.thirteen.datamation.core.spring.DatamationRepository;
 import org.thirteen.datamation.service.DatamationService;
@@ -16,9 +17,12 @@ import org.thirteen.datamation.util.CollectionUtils;
 import org.thirteen.datamation.util.JsonUtil;
 import org.thirteen.datamation.web.PagerResult;
 
+import javax.persistence.EntityManager;
+import javax.persistence.Query;
 import java.util.*;
 
-import static org.thirteen.datamation.core.DmCodes.ID_FIELD;
+import static org.apache.commons.lang3.StringUtils.join;
+import static org.thirteen.datamation.core.DmCodes.DEL_FLAG_DELETE;
 
 /**
  * @author Aaron.Sun
@@ -36,13 +40,18 @@ public class DatamationServiceImpl implements DatamationService {
     }
 
     @Override
+    public void refresh() {
+        datamationRepository.buildDatamationRepository();
+    }
+
+    @Override
     public void insert(String tableCode, Map<String, Object> model) {
         getRepository(tableCode).save(mapToPo(tableCode, model));
     }
 
     @Override
     public void insertAll(String tableCode, List<Map<String, Object>> models) {
-
+        getRepository(tableCode).saveAll(mapToPo(tableCode, models));
     }
 
     @Override
@@ -52,18 +61,73 @@ public class DatamationServiceImpl implements DatamationService {
 
     @Override
     public void updateAll(String tableCode, List<Map<String, Object>> models) {
-
+        getRepository(tableCode).saveAll(mapToPo(tableCode, models));
     }
 
     @SuppressWarnings("unchecked")
     @Override
     public void delete(String tableCode, String id) {
-        getRepository(tableCode).deleteById(id);
+        // 判断是否包含删除标识字段
+        if (containsDeleteFlag(tableCode)) {
+            // 查询语句
+            String sql = String.format("UPDATE %s SET %s = ?1 WHERE %s = ?2",
+                getClassInfo(tableCode).getClassName(),
+                getClassInfo(tableCode).getDeleteFlagField(),
+                getClassInfo(tableCode).getIdField());
+            // 创建实体管理器
+            EntityManager em = datamationRepository.getEntityManagerFactory().createEntityManager();
+            // 开启事务
+            em.getTransaction().begin();
+            createQuery(em, sql, DEL_FLAG_DELETE, id).executeUpdate();
+            // 提交事务
+            em.getTransaction().commit();
+        } else {
+            getRepository(tableCode).deleteById(id);
+        }
     }
 
     @Override
     public void deleteInBatch(String tableCode, List<String> ids) {
-
+        // 查询参数
+        Object[] params;
+        // 查询语句
+        String sql;
+        // 待处理参数的序号
+        int paramIndex;
+        // 判断是否包含删除标识字段
+        if (containsDeleteFlag(tableCode)) {
+            // 查询参数
+            params = new Object[ids.size() + 1];
+            params[0] = DEL_FLAG_DELETE;
+            // 待处理参数序号为2
+            paramIndex = 2;
+            // 查询语句
+            sql = String.format("UPDATE %s SET %s = ?1 WHERE",
+                getClassInfo(tableCode).getClassName(),
+                getClassInfo(tableCode).getDeleteFlagField());
+        } else {
+            // 查询参数
+            params = new Object[ids.size()];
+            // 待处理参数序号为2
+            paramIndex = 1;
+            // 查询语句
+            sql = String.format("DELETE %s WHERE",
+                getClassInfo(tableCode).getClassName());
+        }
+        // 动态拼接查询条件，及查询参数
+        List<String> equations = new ArrayList<>();
+        for (int i = 0; i < ids.size(); i++) {
+            equations.add(String.format(" %s = ?%d", getClassInfo(tableCode).getIdField(), (i + paramIndex)));
+            params[i + paramIndex - 1] = ids.get(i);
+        }
+        sql = sql + join(equations, " OR");
+        // 创建实体管理器
+        EntityManager em = datamationRepository.getEntityManagerFactory().createEntityManager();
+        // 开启事务
+        em.getTransaction().begin();
+        createQuery(em, sql, params).executeUpdate();
+        // 提交事务
+        em.getTransaction().commit();
     }
 
     @SuppressWarnings("unchecked")
@@ -74,7 +138,7 @@ public class DatamationServiceImpl implements DatamationService {
 
     @Override
     public PagerResult<Map<String, Object>> findByIds(String tableCode, List<String> ids) {
-        return findAllBySpecification(tableCode, DmSpecification.of().add(DmCriteria.in(ID_FIELD, ids)));
+        return findAllBySpecification(tableCode, DmSpecification.of().add(DmCriteria.in(getIdField(tableCode), ids)));
     }
 
     @Override
@@ -150,9 +214,58 @@ public class DatamationServiceImpl implements DatamationService {
     }
 
     /**
+     * 由表名获取对应类信息（用于生成po）
+     *
+     * @param tableCode 表名
+     * @return classInfo
+     */
+    private ClassInfo getClassInfo(String tableCode) {
+        return datamationRepository.getPoClassInfo(tableCode);
+    }
+
+    /**
+     * 由表名获取对应实体类的主键字段，驼峰命名形式（不支持联合主键）
+     *
+     * @param tableCode 表名
+     * @return 主键字段，驼峰命名形式
+     */
+    private String getIdField(String tableCode) {
+        return getClassInfo(tableCode).getIdField();
+    }
+
+    /**
+     * 判断是否包含删除标识字段
+     *
+     * @return 是否包含删除标识字段
+     */
+    private boolean containsDeleteFlag(String tableCode) {
+        return getClassInfo(tableCode).containsDeleteFlag();
+    }
+
+    /**
+     * 创建查询对象
+     *
+     * @param em 实体管理器
+     * @param sql sql语句
+     * @param params 参数集合
+     * @return 查询对象
+     */
+    private Query createQuery(EntityManager em, String sql, Object... params) {
+        // 创建查询对象
+        Query query = em.createQuery(sql);
+        int i = 1;
+        while (i <= params.length) {
+            query.setParameter(i, params[i - 1]);
+            i++;
+        }
+        return query;
+    }
+
+    /**
      * map转po对象
      *
      * @param tableCode 表名
+     * @param map map对象
      * @return po class
      */
     private Object mapToPo(String tableCode, Map<String, Object> map) {
@@ -164,9 +277,10 @@ public class DatamationServiceImpl implements DatamationService {
      * map转po对象
      *
      * @param tableCode 表名
-     * @return po class
+     * @param maps map对象集合
+     * @return po class集合
      */
-    private Object mapToPo(String tableCode, List<Map<String, Object>> maps) {
+    private List<Object> mapToPo(String tableCode, List<Map<String, Object>> maps) {
         List<Object> objects = new ArrayList<>();
         Class<?> poClass = datamationRepository.getPoClass(tableCode);
         for (Map<String, Object> map : maps) {
